requirements.txt
pillow
jira
